support streaming from methods that don't have __iter__ implemented
httpx requires the content to be an iterable (excluding str, bytes).
The file object may not always have __iter__ implemented.

This could happen, for example, when the file object is wrapped
(eg: CallbackIOWrapper of tqdm) using __setattr__ hacks, but the
wrapper usually don't implement the __iter__ methods.

__iter__ method in python is a special method which should exist on the
class itself and otherwise might end up throwing TypeError during iteration.

I have hacked that around by calling read() methods in that case and manually
reading through in that case. I know, this might not be ideal, but the only choice
I have is to open a feature request to httpx or in tqdm, for which I am not sure right now.