optimize: improve upload speed by using larger chunks
I noticed that httpx was using __iter__, so we were sending a
a newline as a chunk rather than a large chunk. I am not sure
if this is something they want to have upstream, though I might
start discussion there. For now, with this chunking, we are at
the same performance as using requests