retry/reconnect on connection failure
We do this on the basis of feature detection. We try to send
a OPTIONS request to the base url and see if it advertises
'Accept-Ranges' header. Otherwise, we just use the header from
the opened network request. This is done in this way because
ownCloud/Nextcloud does not advertise Accept-Ranges header on
GET requests.