client: handle propfind response from Nginx dav ext
It sends a 207 MULTISTATUS response and contains a 404 inside
propstat:status instead of just returning a 404 status code.

It's a bit tricky to use propstat:status for webdav4, as they may
be status for properties, not the resource, and might impact other
webdav servers like Owncloud/Nextcloud, as they return multiple
propstats.

For now, the change only affects propfind users. Also fixed
`isfile` and `isdir` to not raise ResourceNotFound error, and
instead just return False.