update httpx to 0.20.0 and pin it
This is likely going to be the last release before 1.0 release
for httpx, so pinning it. Will unpin after 1.0 release.
New version of httpx stopped following redirects by default.
I am not really sure how much it'll impact, so I'm choosing to keep
it to the same defaults as httpx.